azure-vnet-simple-vm/python-worker-app/requirements.txt
numpy
pandas
opencensus-ext-azure